fix(terminal-plugin,vnc-plugin): batch W — ensureMap guard, no-console, VncHost type (#8417 #8418 #8419)
GH#8417: add ensureCommandHistoryMap() guard in useTerminalStore — prevents
crash when Pinia rehydrates commandHistory as plain object instead of Map.
Mirrors the existing ensureMap() pattern for sessions.

GH#8418: replace console.* with noop in autobot-plugins/terminal/src/utils.ts
and autobot-plugins/vnc/src/utils.ts — no-console rule compliance. Standalone
plugins cannot import from the host app's debugUtils.

GH#8419: move VncHost interface from VncViewer.vue SFC to
autobot-plugins/vnc/src/types.ts; update VncViewer.vue to import it;
update vnc/index.ts to re-export from ./src/types.
