+++ autobot-plugins/vnc/index.ts
@@ -9,7 +9,7 @@ import VncToolbar from './src/components/VncToolbar.vue'
 export { VncViewer, VncToolbar }
 export { useVncControls } from './src/composables/useVncControls'
 export type { VncActionResponse, MouseClickParams, MouseDragParams, MouseScrollParams } from './src/composables/useVncControls'
-export type { VncHost } from './src/components/VncViewer.vue'
+export type { VncHost } from './src/types'
 
 export const VncPlugin = {
   install(app: App) {
